Ingredient Part
I did the create, edit, update and delete part but I’m really not sure
it it’s the right way to do it. It was the first time, I used html, css
and angular js so I inspired myself with the company code ! I still
have to work on it and I’m still working on the part for the list of
ingredient with the gastronomist.